changed order of header files

+++ 0x0E-function_pointers/1-array_iterator.c
@@ -1,6 +1,7 @@
+#include <stdlib.h>
 #include "function_pointers.h"
 #include <stdio.h>
-#include <stdlib.h>
+
 
 /**
  * array_iterator - executes a parameter on each element of an array
@@ -14,7 +15,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i = 0;
+	unsigned int i = 0;
 
 	while (i < size)
 	{
